Section 5, Lecture 226: AI Hearing Perception and Make Report
In NPC_AI_BP
Edit the AI Sight config of the Radial and the VisionAngle.
Add one for AI Hearing , Same as the Last Lecture 225. Set the Hearing Range and Affiliation.

In FirstPersonCharacter BP
Call On Landed Event ---> Report Noise Event
Location we need to get from the actor (self)
Instigator is who (self)

See more on In NPC_AI_BP and FirstPersonCharacter BP