Section 5, Lecture 259: AI can shooting
Is continue the last Lecture 258, more BP code are related to.

Do a Additive Animation for it. Duplicate the Fire Rifle and rename it. Set it Additive as last time we learn.

Hero BP is connect Identify Pose and Additive

PatrollingGuard for the variable connect to

NPC AI BP for when AI will shoot

Please refer those of the edited File
tips by using Unreal diffr